#
Mejoras en la gestión de reservas y refactorización

Se ha cambiado el tipo de la propiedad `NumeroHuespedes` de `sbyte` a `int` en `ActualizarReservaRequest.cs` y `Reservas.cs`, permitiendo un rango más amplio de valores.

Se ha modificado el método `ObtenerReservasPorClienteIdAsync` en `ReservaApplicationService.cs` para utilizar la sintaxis de null-coalescing (`??`), mejorando la legibilidad y el manejo de excepciones.

En `CrearReservaAsync`, se ha añadido una verificación para cambios en fechas o categoría de habitación, optimizando la gestión de disponibilidad.

Se han eliminado métodos en `ServicioApplicationService`, sugiriendo una refactorización de funcionalidades no utilizadas.

Se ha añadido un parámetro opcional `idReservaActual` al método `HayDisponibilidadAsync` en `IReservaRepository` y `ReservaRepository`, mejorando la lógica de actualización de reservas.

Se han realizado mejoras en el manejo de excepciones, utilizando la sintaxis de null-coalescing para lanzar excepciones más claras.

+++ SGHR.Application/DTOs/Reservas/ActualizarReservaRequest.cs
@@ -21,7 +21,7 @@ namespace SGHR.Application.DTOs.Reservas
         public EstadoReserva Estado { get; set; }
         
         [Required(ErrorMessage = "El numero de huespedes es requerido")]
-        [Range(1, sbyte.MaxValue, ErrorMessage = "El numero de huespedes debe ser al menos 1 y no exceder 255")]
+        [Range(1, int.MaxValue, ErrorMessage = "El numero de huespedes debe ser al menos 1 y no exceder 255")]
         public sbyte NumeroHuespedes { get; set; }
 
     }
